+++ WebbLabb2/Shared/DTOs/ProductDto.cs
@@ -6,7 +6,7 @@ public class ProductDto
     public int ProductNumber { get; set; }
     public string ProductName { get; set; } = string.Empty;
     public string ProductDescription { get; set; } = string.Empty;
-    public int ProductPrice { get; set; }
+    public float ProductPrice { get; set; }
     public string ProductCategory { get; set; } = string.Empty;
     public int OrderCount { get; set; }
     public bool ProductStatus { get; set; }
